ledctl/requirements.txt
rpi.gpio
adafruit-blinka
adafruit-circuitpython-neopixel
redis